Make LQTYStaking.stake revert with zero amount
Before, stake and unstake functions of LQTYStaking wasn’t checking if
the amount passed as a parameter was zero. As they both send gains to
the owner, and there’s no other function for that purpose, at least
one of them should allow the amount being zero.

For consistency with StabilityPool, the fix applied is:

- Revert in stake
- Allow it in unstake, but skip the lines that don’t make sense with a zero amount.