Fix DefaultPool / pending rewards bug in liquidations
`_liquidateRecoveryMode` moves the trove's rewards from DefaultPool -> ActivePool, but in the final branch, does nothing with the trove, and does not update its coll/debt/snapshots.

The trove is left with pending rewards (as determibned by the reward snapshots on its struct),however, its actual allocated ETH and LUSDDebt rewards have been moved from the DefaultPool -> ActivePool. Later when the trove is included in another liquidation (or potentially even other operations), the excess pending rewards recorded on the trove can cause the transaction to revert, as there may not be enough in the DefaultPool.

Fix:

In `_liquidateRecoveryMode`, only move rewards from DefaultPool -> ActivePool in the cases where the trove gets closed. Do nothing with rewards in the final branch of `_liquidateRecoveryMode` in which the trove is not closed or changed.

Signed-off-by: Rick <richkws@gmail.com>