Implement minimum net debt
In a redemption sequence, when the partial hint is out of date, we return 0 for the trove's LUSDLot.

The redemption loop breaks when it finds a trove with 0 LUSDLot:

`if (singleRedemption.LUSDLot == 0) break;`

However, there can exist troves with no net debt (only the 50 LUSD liquidation reserve). These "0 net debt" troves cause the redemption loop to break, and could even constitute a DoS attack vector - if there are troves with 0 net debt at close to 110% ICR, they can cause a redeemCollateral transaction to break early, redeeming little or even no LUSD. Such troves could easily be opened and kept at close to 110% ICR.

More generally, with small troves at the ICR = 110% boundary, redemptions can be unprofitable. The smaller the minimum net debt, the lower the gas price & ETH price at which redemptions can be unprofitable.

Fix:

- Implement a minimum net debt. Provisionally: 1950 LUSD (to ensure 2000 LUSD total debt)
- Make redeemCollateral not perform a partial redemption that would cause the trove's remaining net debt to be less than the minimum net debt.
- Use a bool to signal whether the partial hint got out of date (or not), and break the loop conditionally on that

HintHelpers `getRedemptionHints` is unchanged in order to ensure we're really testing that the core contracts skip the partial if its resultant net debt would be above the minimum.

There's no real advantage (AFAIK) to having any extra computation in
`HintHelpers` to calculate whether or not the partial will occur, though
it could be done easily.

Signed-off-by: Rick <richkws@gmail.com>